services/pulsecast_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
pydantic-settings
kafka-python
requests
pytest
httpx
